Update patch set 2
Patch Set 2:

(5 comments)

> I used cat command to test code. When I tried to use the size of file as buffer size then it gave me a error "*** stack smashing detected ***" in read_pos function. So, I kept the buffer size as _length and now it reads only 10 characters of a file.

I don't understand which buffer you refer to. The design of the filesystem is:

"cat" (or anything that reads from disk) will allocate its own buffer. It will call the read() or fread() function passing the buffer and length. Eventually this calls ufs2_read.

The job of ufs2_read is to read no more than the passed length (it is allowed to read less, but usually it will fill the buffer).

ufs2_read() will be called again with a different position in the file, and should read more bytes.

You have ignored both parameters (the position and size of the read) in your code. So, indeed you are probably writing more bytes than there is space in the buffer, and this ends up crashing.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>