Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

> Patch Set 1: Code-Review-1
> 
> This seems like an extremely obscure layout. I am not sure we really want or need to maintain it in the tree...

There is nothing to maintain here. We add the file and we never touch it again.
Also, it's not even the first 'obscure' layout, we already have the Kinesis ones and that's not really a common keyboard either.

I think it's fine to use a few bytes of data in the Haiku package for this.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2