Update patch set 1
Patch Set 1: Code-Review-1

As ARM is a new Haiku platform, do we even need ctor/dtor. The GCC info seems to suggest that it is preferred. Also I guess we should add it for x86 platforms. There is also a problem with init order but maybe only be when using a gcc extention(?): ctor should be done in one direction and init_array in the other.

So while this might be ok'ish, I think we should either have answers to above, or have comment in the linker script about the above.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1