Update patch set 1
Patch Set 1:

By reading the gcc ticket, it seems that even gcc developers are pretty unclear about this.

For the bootloader, I wouldn't care too much. We are not doing anything too crazy (such as using ghe gcc extensions to set constructors and destructors priorities) and no one else is going to link code using that script. So, if it works for running the bootloader, it's goog enough.

C++ says the order in which constructors are called is undefined, and the main concern in the gcc ticket is about what happens when linking static libraries, to make sure the libraries are initialized before the main code.

I think we should keep things simple for the bootloader, and avoid implementing all quirks and legacy of the standard way here, unlike in the runtime_loader where we have to do things as correctly as possible.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>