Update patch set 1
Patch Set 1:

We need to rebuild the repository on the haiku server rather than merging this. Just taking note here (I think only waddlesplash knows the process for this).

I tested that the system builds and boots with this (but on 32bit that doesn't test much, because it's only the secondary arch).

Then changes I commited are backwards compatible with the currently used version of ICU.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>