Conversion error while communication Input Server with InterfaceDefs
I am trying to add the BString in the InterfaceDefs.cpp and trying to reteive it in the InputServer by printing the logs in the syslog file. But, I am getting the output of the BString in the binary format.

For example, In the line 9228 and 9229, The BString is 'ABC' which gets printed in the logs from the InterfaceDefs file.
But, in the line 9230, the logs are printed in the Input Server, the should should be similar to the above lines but, its now in the binary format.

Please let me know if I am doing wrong somewhere!

https://imgur.com/5wmfctx

Change-Id: I3dd51fd174726eb7c24b4d54fc5e65322b7c6ed0