Update patch set 5
Patch Set 5:

> > uint32 _reserved[6 - sizeof(BRect*)/sizeof(uint32)];
> 
> I tried this and checked that the BTextView class size is 512 bytes on 64-bit with or without this change. I assume there must be some built-in padding on 64-bit making this change unnecessary.

Where would the padding come from? The change is needed unless oyu can explain how things are laid out in memory and why it doesn't make a difference.

I don't see where "built-in" would come from. The compiler is not magically deciding to add random padding for no reason?

One possibility is that the class size is rounded up to 64 bytes, but then, not changing this only makes things even more confusing for the next person changing the class, because they will have to understand that the padding is left intentionally at the wrong size and relying on the compiler to make up for it.

I think I would go with the simplest solution: don't use a pointer, just store the rect normally. It's 4 floats so there is space for it. We can always reorganize the data later if we run out of space.

Patch-set: 5
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>