Installer: fix 'clean install' over existing installation.
Whenever the target of an installation is a partition that is not empty, the
Installer prompts the user whether they would like to continue with a 'clean
installation', i.e. an installation that clears out the system folder,
excluding the settings, and cleanly installs the assets in the source's
system folder. At all other locations the source data is merged, meaning that
the source version is copied in place.

The logic that clears out the existing /system/ directory stopped working. This
commit changes the logic in such a way that existing directories under the
/system/ directory will be clobbered (i.e. removed) except for those that are
in /system/settings and below.

Furthermore, there is a change in the logic that makes sure that if the target's
directory was clobbered, that it will be recreated.

The change was manually tested.

Fixes #16092