Appearance: Save decorator path instead of name.
Make fCurrentDecor and fSavedDecor point to the decorator path
instead of the name.

Decorator name has to be turned into a path before saving to disk. It
is better to set the path we have then to have DecorInfo look for the
name in the four possible locations it could be in to find the path.

This matches ControlLook which already uses the path.

Note that the path used for the default decorator is simply "Default".