Update patch set 2
Patch Set 2:

this patch is required to support Haiku in proxychains-ng.
note that for correctness, the variable name of "int type" shouldn't be changed, and the FIXME comment can be removed, as const void* is the right type and can be safely cast to u_char*, as is done just the next line already.

Patch-set: 2
CC: Gerrit User 1000246 <1000246@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>