Update patch set 5
Patch Set 5:

The pull request was somehow missing the change of the implementation, so I've taken the liberty to merge the commit from the pull request with the original patch linked. I've also corrected a pointer arithmetic warning that was introduced by the change (by removing the unneeded line).

There is no need for "tricks" with the email address. The push was probably rejected because of the unverified user, not because of the commits it tried to push. Pushing changes authored by someone else, with addresses not even known to Gerrit, is quite usual and doesn't need any workarounds.

Admittedly the setup here is somewhat inconvenient as it forces the use of a Github account. You did the next best thing and opened a pull request at the Github mirror, which is generally fine IMO, as the commit can be moved over to Gerrit reasonably easily. It's likely that such pull requests are sometimes missed as the mirror isn't actively monitored by many people. In this particular case the PR was also closed with a pretty terse comment, which wasn't too helpful or nice, so thanks for following up anyway.

I've verified the build for both x86_gcc2 and x86_64, so if the modified commit is fine by you this can be merged.

Patch-set: 5
Reviewer: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>