Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This doesn't sound like a good idea to me. At least the general idea behind storing the name only is that you can override system add-ons with user add-ons.
> Or am I missing something?

If that is the intention of the code I was unaware of it. I just figured that there was no compelling reason to turn the name into a path on save when we already have the path.

This is more cleanup done along the way towards fixing #16412. If undesired this change can be abandoned.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>