Update patch set 1
Patch Set 1:

> Patch Set 1:
> If that is the intention of the code I was unaware of it. I just figured that there was no compelling reason to turn the name into a path on save when we already have the path.

There is definitely no compelling reason to do that on save. The general idea with add-ons is to store only the name, and then look it up on load on all places in a fixed order -- this is very similar to how the $PATH variable is supposed to work. It's quite possible, that not all code behaves this way, though.

But if this code stored the complete path anyway, your changes definitely improve the situation.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>