Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Some details regarding the segmentation fault:
> So what I noticed is after I call the command "cat -n 'numberOfBytes' file.txt", with less number of bytes (Have tried 100,000 bytes) it works okay. But on bumping it up to high number of bytes, like 20MB, I get a segmentation fault.
> And after tracing I found that it happens after xfs_open() is finished executing and before xfs_read() hook starts executing. I couldn't figure out which calls are made in between the calling of these two hooks. But I am sure it's before xfs_read is called.

Both the 10 byte issue and the segfault are the fault of how https://git.haiku-os.org/haiku/tree/src/tools/fs_shell/fssh.cpp#n692 is implemented. By default it explicitly tells cat to only read 10 bytes even without a -n specifier. Furthermore, it allocates the output buffer on the stack, which leads to a segmentation fault due to stack overflow if too large of a size if specified. This should probably be altered so neither of these are the case any more (or we should move to further fleshing out the FS from within Haiku via userlandfs upon reaching a point like this).

Patch-set: 1
CC: Gerrit User 1000015 <1000015@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>