Update patch set 1
Patch Set 1:

As a sidenote, I am not sure whether "Important notice" is a good idea for a title, or whether I should either revert it or use something different instead.

Also, would it be a good idea for me to share screenshots of small visual changes, in order for more people to be able to chime in easily?

Patch-set: 1
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>