Update patch set 1
Patch Set 1:

I would like something more visual special than a pop up menu. I was thinking of a second tab with a grid of rectangular radio buttons showing where Deskbar would go in each mode.

Unfortunately one of the reasons the drop down isn’t working is probably because I have overloaded the corners. You have to pass a vertical bool parameter for the corners to decide if which of the two mini mode they are in. The BDeskbar API may need to be extended too. :/

Patch-set: 1
CC: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>