Update patch set 2
Patch Set 2:

I do look forward to a nice thumbnail view. Thanks for working on it!

I tried out patch 1 and regret to say that I see performance issues. When entering a folder with ca. 500 items, it takes quite a while to generate the thumbnails. Moving the scroll bar becomes unreactive. Even when closing and re-opening the folder, the scrolling is not smooth.

Switching to List view takes over a second and the mouse pointer is leaving a white trail when moving over the file icons.

Maybe things can speed up with multiple threads, as I see only one core being used?

Patch-set: 2
CC: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>