Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > EULA window should remain a modal window.
> > 
> > I also find the current modal windows unintuitive since they do not have tab titles, but that's a different topic.
> 
> I seriously do not think that showing a modal window immediately after a titled window is good for new pairs of eyes, especially under the pretext that people usually look at window titles in order to figure out what's going on. I'll disagree with you in that regard, because I don't think it is irrelevant-- after all, this is the fact that resulted in my willingness to make that small change.

The functional case for making the window this way is that we want it to stand out. Like an eyesore. It is specifically to designed to make a user review the contents of the window. So there is a functional reason for having this Window differ visually.

Patch-set: 1
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>