Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (1 comment)
> 
> SetMouseSepeed
> should be
> SetMouseSpeed
> check for any other misspellings like this

I will be careful next time to avoid these type of mistakes.

> How do you know the settings are being saved and reloaded?  Try writing a small simple program which saves and loads the settings and verify you get back what you saved. Maybe have the test program allow command-line options to set specific settings so that you can verify they are changing.  Sometimes taking a step back and testing and verifying your assumptions can save you from mistakes further along.

I have printed the logs to check whether the settings are stored and retrieved or not. From the frontend, the GUI is working properly. I think I have to make a simple program if the settings are not working. But, here the problem is the settings are working but just for one mouse and get applied to all other mice. According to me, the problem is in the InputServer under the HandleGetSetMouseSpeed I am getting the name field empty! But under the Input preferences code, everything is working fine.
Or, I get the name of the devices in the server-side somehow, then it is in the binary format! 
 
> Myself I would have used the "device ID" instead of the "device name", as with the ID it's always the same sized variable, and is unique to a specific model of device.
> You might even switch to saving plain human readable text data, until you are satisfied that you and saving and reading back the expected data.  That way you can open the settings_file with a text editor and see whats in it.

For now, I chose the device name to store the settings. Once it gets working I will add the device ID instead of the device name.

> Also be sure to heed Axeld's comments above and fix function names, variable names, etc to comply with the Haiku coding guidelines.

Will do that.

Patch-set: 3
Reviewer: Gerrit User 1000116 <1000116@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>