Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I don't think this change is necessary.
> 
> These terms do not have racial origin, and looking for racism where there is none achieves nothing.

I am with you that the terms itself do not come from concepts with a racial origin, but history should not be the standard of measure for anythings. Words and expressions gain and lose meaning continuously; measuring everything by the 'original' meaning will ultimately lead to endless bickering between two parties that are literally stuck in two different time periods.

Unfortunately for all the blacklists and whitelists out there, the metaphor they use (black = bad and white = good) overlaps with a contentious (American-centric?) debate about racism.

> This is an empty gesture and IMO an example of cargo culting.

Would I have made/proposed this change myself? Probably not. However, there's enough in this change that will give it the benefit of the doubt. First of all it replaces a metaphor with a semantically meaningful alternative (black will become block). Secondly, it is backward compatible.

Patch-set: 2
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>