Update patch set 2
Patch Set 2:

> Patch Set 2:
> I am with you that the terms itself do not come from concepts with a racial origin, but history should not be the standard of measure for anythings. Words and expressions gain and lose meaning continuously; measuring everything by the 'original' meaning will ultimately lead to endless bickering between two parties that are literally stuck in two different time periods.

Yes, this is called euphemism treadmill. The issue I have with this is that everything can be perceived as problematic. "Flip" is a pejorative for Filipino. Should we stop "flipping bits"? Maybe it's time to purge the codebase of this very clearly offensive term?

Context matters.

> Unfortunately for all the blacklists and whitelists out there, the metaphor they use (black = bad and white = good) overlaps with a contentious (American-centric?) debate about racism.

This metaphor is very deeply ingrained in Western culture though. I really don't want to generalize because I'm not an expert on cultures, but it appears that this kind of dualism and association is pretty widespread elsewhere.

It's just how the world is. Light = good; dark = bad, fear, death (rot is black).

Are we going to rewrite languages because of this? What about "black magic"? "Black markets"? "Black hats" and "white hats"?

One solution could be to stop calling people "white" and "black" - which are pretty inaccurate descriptions. This is something I could get behind.

> Would I have made/proposed this change myself? Probably not. However, there's enough in this change that will give it the benefit of the doubt. First of all it replaces a metaphor with a semantically meaningful alternative (black will become block).

I agree. That would have been a good reason. However, the intention stated is to be "inclusive". I don't buy that. This does nothing to help minorities. It doesn't stop racism, because the use is not racist.

Anyway, I'm not hell bent on not having this change merged (that's why I gave it -1, not -2). But I do think voicing my objection to such unwarranted language policing is important.

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>