Update patch set 2
Patch Set 2:

> Patch Set 2:
> > Would I have made/proposed this change myself? Probably not. However, there's enough in this change that will give it the benefit of the doubt. First of all it replaces a metaphor with a semantically meaningful alternative (black will become block).
> 
> I agree. That would have been a good reason. However, the intention stated is to be "inclusive". I don't buy that. This does nothing to help minorities. It doesn't stop racism, because the use is not racist.
> 
> Anyway, I'm not hell bent on not having this change merged (that's why I gave it -1, not -2). But I do think voicing my objection to such unwarranted language policing is important.

I think the author of the change made a smart choice by stating the background neutrally (i.e. "OSS projects are doing this" and not "Haiku should be more inclusive by removing language"), so to me it passes the smell test of things getting too political.

Will not upvote before language in commit message is fixed though ;-)

Patch-set: 2