Update patch set 8
Patch Set 8:

> Did not modify the 32/64 bit time_t workaround as the suggested solution does not address the root problem. Anyone got a better idea on how to address comparing time_t of different sizes values across architectures?

I would always store them as 64bit. This makes the on-disk storage future proof and usable the same way on all architectures.

On x86 32bit you then cast it to 32bit to compare it with a time_t, this will work ok at least until 2038. By then I hope R1 will be released and we will have switched to 64bit time_t in R2.

This means the workaround is only in the R1 32bit x86 case, and we don't have to worry about it in other architectures. It does not have any drawback (that architecture is going to fail in 2038 no matter what you do here), and it keeps the consequences to a minimum (no duplicate thumbnails generation, no attribute format conversion will be needed)

Patch-set: 8
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>