Update patch set 2
Patch Set 2:

A lot of debate over a single-character change (as expected?).

Yes, surely this change will not help fix racism. It may not be worth our efforts to track every single use of these terms and replace them. But if someone is doing the work and sending patches, I see no reason to reject them either.

"blocklist" has a clearer meaning, sowe're not making things less understandable here. You can see in the wikipedia article about "blacklist" that various tech companies are making this change as well (github, apple, linux, redhat…) so it seems quite possible that the term will fall out of use.

Anyway, I'd say the best thing to do is not waste time debating this and accept the patch, now that the work is done (with my comments about backwards compatibility addressed). I see no downside to it, at least.

Patch-set: 2