Update patch set 2
Patch Set 2:

> "blocklist" has a clearer meaning, sowe're not making things less understandable here.

As I've said before, fair enough.

> various tech companies are making this change as well (github, apple, linux, redhat…)

And we should follow their whims because...?

> so it seems quite possible that the term will fall out of use.

We could revisit the terminology when it does.

> Anyway, I'd say the best thing to do is not waste time debating this

Very convenient:

* If I agree/don't care debate is a waste of time.
* If I disagree we need to debate.

Not a very productive attitude.

To make my position clear:

I'm OK with accepting this based on "previous word was unclear [was it though?], we should make it clearer".

I'm not OK with accepting based on "others do this too" (which is a very weak argument and would not fly with technical changes - compatibility excluded) or "inherent racism/inclusivity".

Maybe it's living in a country where institutional censorship was implemented in the past, that makes such moves rub me the wrong way.

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>