Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > "blocklist" has a clearer meaning, sowe're not making things less understandable here.
> 
> As I've said before, fair enough.
> 
> > various tech companies are making this change as well (github, apple, linux, redhat…)
> 
> And we should follow their whims because...?
> We could revisit the terminology when it does [fall out of use].

Sure. The patch is there, so we may as well do it now.

> 
> > Anyway, I'd say the best thing to do is not waste time debating this
> 
> Very convenient:
> 
> * If I agree/don't care debate is a waste of time.
> * If I disagree we need to debate.

That's not what I meant, sorry I was not very clear. We could debate this as much as we want, we both already admitted that we don't care that much about it. I won't be actively making such changes, but if someone does, I think it's ok to merge them.

We could debate wether changing the wording is a good or bad thing, but I'm not really interested in that. So, I said I would be wasting my time in doing so. Not because I have strong opinions one way or another, just because in the end it doesn't really matter to me. I understand the meaning either way, I'm not particularly annoyed by one term or the other.

If people want to debate this and feel strongly either way, I'll let them discuss, and probably accept the desision either way.

> 
> Not a very productive attitude.
> 
> To make my position clear:
> 
> I'm OK with accepting this based on "previous word was unclear [was it though?], we should make it clearer".
> 
> I'm not OK with accepting based on "others do this too" (which is a very weak argument and would not fly with technical changes - compatibility excluded) or "inherent racism/inclusivity".

I think more in terms of "why would we reject this change?" (not specifcally for this change, but in general when I review code, my mindset is that way) and the answer would be something in the line of "we don't want to get into an evolution of the English language which was started by tech companiess". And that raises more questions for me: why don't we? It doesn't cost us anything and I don't see how it can hurt (but I could be missing something), especially considering we have an alternative that is easily understandable.

So, I don't see reasons to reject, but I don't necessarily see reasons to accept, either. Since someone else did the work already, that's enough for me to lean on the "ok, fair enough" side and accept the change.

> Maybe it's living in a country where institutional censorship was implemented in the past, that makes such moves rub me the wrong way.

I don't know, not being a native English speaker nor a linguist I probably fail to see some of the nuance and connotations, knowing the term only in its software engineering usage. So, we probably all have different relationships to English and to our other languages, for all kind of cultural reasons.

To me, no meaning is lost by replacing "blacklist" with "blocklist" (on the contrary, the latter is more descriptive). I wouldn't be ok with using a less clear word (that would be a resons to reject the change), but here, I think it's fine. The improvement may be small or non-existent to us, but I don't have anything against it, and it was apparently annoying enough for someone to put the effort in creating this change request.

Anyway, back to coding now :)

Patch-set: 2