Update patch set 1
Patch Set 1:

Well, first of all, the window is misnamed, it is not an EULA.

And that leads to the deeper issue: no one reads EULA. So, putting important information here is a bad choice, and happened only for historical reasons (BeOS did have an EULA window and we just blindly copied it at the time).

The result isn't great, this window has some important information for the setup process but there is no way to keep it around, print the thing, copy it to somewhere else.

Should it have a title? If it were an EULA, I would say no: the point of the EULA being modal is to highlight the fact that it is a blocking thing: you must read the EULA before continuing. But, it is not. Is the information here important enough that we want to push users into reading it in the same way? Yes, but it doesn't work because users are well trained in skipping such screens. So now we have to resort to another trick (additionally or in replacement). Maybe some better formatting of the text and inclusion of some sreenshots would better grab the user attention, for example?

In any case, just removing the modal state doesn't feel right.

As Emir said, there is the separate question of wether modal windows should have a titlebar. If we decide they should, this should be fixed at the OS level, not by removing all modal windows. But then I agree with nielx, it seems appropriate that they don't, and that they are really blocking you out from doing other things by being there in the middle of the screen. We know that's annoying, but it's on purpose. If it feels strange, it's possibly because a particular window shouldn't be modal in the first place. But it can also be because that design choice is specific to BeOS and Haiku and unexpected when coming from other OS. We should think about which of these two options is playing here (or possibly the accumulation of both of them)

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>