Update patch set 8
Patch Set 8:

At least in packagefs, we used SYS: as a prefix rather than HAIKU: for system attributes. This makes them independant of the OS name, in case someone forks and renames Haiku, or if someone else decides to use bfs attributes in their OS or whatever.

For the year 2038 bug: storing the attribute as 32bit does not fix the problem either. So let's at least have a single format, and one that can store dates farin the future. There are workarounds to keep things working for some more time, for example casting time_t to unsigned extends its range by another 40 years or so. But I hope our gcc2 compatibility will be gone in 2038 and we don't even need that?

Patch-set: 8