Update patch set 1
Patch Set 1:

(3 comments)

As the results of a choice are instantly visible, contrary to the active corners of the Screensaver prefs, I'm not convinced a graphical representation has much merrit. Also, the different mode-options may be difficult to accomplish in such a "miniaturized" representation. For the user it may be more difficult to discover and snap into the different modes, than doing the real thing by dragging the knobbly bits of the Deskbar tray.

Maybe sub-menus can help:

- Top bar
- Bottom bar
- Top right
     |
     --- expanded
     --- mini horizontal
     --- mini vertical
- Top left
     |
     --- expanded
     --- mini horizontal
     --- mini vertical

...etc.

I'd put the traditional "Top right - expanded" first, of course. :)

Patch-set: 1
CC: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>