Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Users are well trained in skipping such screens. So now we have to resort to another trick (additionally or in replacement). Maybe some better formatting of the text and inclusion of some sreenshots would better grab the user attention, for example?
> 
> In any case, just removing the modal state doesn't feel right.
> 
> We know that's annoying, but it's on purpose. If it feels strange, it's possibly because a particular window shouldn't be modal in the first place. But it can also be because that design choice is specific to BeOS and Haiku and unexpected when coming from other OS. We should think about which of these two options is playing here (or possibly the accumulation of both of them)

I know that the annoyance was caused on purpose, but my question was as to whether this annoyance actively prevents the behaviour that I'm also concerned about, or reinforces it. Additionally, I reluctantly opted for "Important notice" (rather than "EULA"), after tracking down the changes and the aforementioned origin, and, well, the fact that it wasn't an EULA, like you just mentioned.

Either way, I've been poking at the Installer for a bit and have had a LOT of other changes that I've been meaning to implement lately. This is just the smallest possible step that I could take for the time being, so I just decided to take it, but I'm really now beginning to rethink my perspective on the ideas I've had (out of the scope for this change). Is the IRC the best place to discuss this, so I'll just start gradually asking the mailing list for opinions and some additional perspectives. Thanks!

Patch-set: 1
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>