Update patch set 1
Patch Set 1:

A visible representation is always better. With this too many options it is probable that the user will get annoyed by having to try all of them. Definitely bad UX and lazy design. User should have an idea of what is a vertical mini mode and horizontal mini mode without having to change their actual desktop settings. Add the expanded bar mode to that as well.

Patch-set: 1
CC: Gerrit User 1000169 <1000169@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>