Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > 
> > With this too many options it is probable that the user will get annoyed by having to try all of them.
> 
> They don't. Once having discovered what mode is "bar", "expanded", "horizontal/vertical mini" they know their permutations (top/bottom, left/right).

My complaint is that the linguistic descriptions are not as good as visual ones relating to the size and position of Deskbar. A pop up box is minimally acceptable but not ideal.

I was hoping humdinger could get on board here and supply us with one of his (in)famous mockups!

> Having a graphical representation of your screen with a Deskbar homunculus that the user can move around seems to be worse. The user has to blindly move the thing around to accidentally find the different modes.

They would not be blind because the visual would aid their understanding. The current situation is blindness because the dots are invisible to some users. My forum post where I brought receipts triggered this patch set and discussion  to happen I’m afraid.

Click on the box where you want Deskbar to go, and it goes there and gives feedback by marking the appropriate radio button control in preferences.

> 
> > User should have an idea of what is a vertical mini mode and horizontal mini mode without having to change their actual desktop settings.

A visual representation would give you a better understanding of what the different modes mean before switching to them than a pop up menu would.

> 
> The user can try different modes and immediately sees the changes, just like with the rest of the Deskbar settings. The "Revert" button brings back the original setting. Having a sort of preview inside the preference panel would go against this paradigm and would then need an "OK" button to apply the setting.

This is untrue, the update could happen live and the Defaults and Revert buttons could also happen live with a visual representation.

Patch-set: 1