Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Note to self: similar solution (static lib catkeys) should be applied in src/add-ons/translators/shared/Jamfile

I'm not sure about that. If I understand correctly, this results in duplicating the strings in all catalogs for apps that link with the static library. Which is not great, it wastes space and requires more work from the people doing translations. Am I right?

The laternative would be to make a separate catalog for the static library, and not using the B_TRANSLATE macros but using the BCatalog API directly instead to load and access it (the macros use the localestub to automatically find a catalog according to the shared object mimetype, but direct usage of BCatalog allows accessing other catalogs if needed).

It seems fine to use your work for Cortex because there is very limited reuse (or not at all?) of the static libraries. But for the shared translator code, it would result in a dozen copies of the same strings.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>