Update patch set 5
Patch Set 5:

> https://paste.ubuntu.com/p/yx9Hg48ZT5/
> 

This crash is in a call to syslog() in MultipleMouseSettings::GetMouseSettings.

One possible reason for this is that you forgot to use .String() when calling syslog() to log a BString, or a mismatch between the %characters in the log string and the parameters you passed.

The simplest way to avoid this crash is to remove the call to syslog. Of course, checking them and finding the one which causes the problem, is better.

Patch-set: 5
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>