Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I'm not sure about that. If I understand correctly, this results in duplicating the strings in all catalogs for apps that link with the static library. Which is not great, it wastes space and requires more work from the people doing translations. Am I right?
> 
> Yes.
>  
> > The laternative would be to make a separate catalog for the static library, and not using the B_TRANSLATE macros but using the BCatalog API directly instead to load and access it (the macros use the localestub to automatically find a catalog according to the shared object mimetype, but direct usage of BCatalog allows accessing other catalogs if needed).
> 
> Is it possible to link multiple catkeys with different MIMEs into one executable? Maybe I could link multiple catalogs into an executable and then query for them by signature?
> 
> I don't see how I could use BCatalog API to load a catalog from file. There is only one constructor, and it expects entry_ref to executable. I could use MutableLocaleRoster to create CatalogData and load it that way but it is private and seems to be a pretty complicated way for such a simple use case.

Indeed the API does not expose the needed things. There is no technical reason for that, however. The catalogs are not linked into the executable, they are just stored in /system/data/catalogs with a path matching the MIME type.

You can load them using the DefaultCatalog class (in headers/private/locale) but adding something to the LocaleRoster and/or BCatalog to do it more directly is probably a good idea (a constructor taking a MIME type instead of an entry ref would do).

What will be lost is the ability to load a catalog that would be embedded into the executable, if we do it this way. We can go further and have a constructor that takes both an entry_ref and a resource identifier (name) and loads the catalog either from the BResource in the linked ref, or from /system/data/locale/catalogs if the resource does not exist (as is the case for the normal catalog loading).

Finally, for Cortex there is the option of just putting all these strings in the Cortex app catalog, add-ons could explicitly refer to it if they can get an entry_ref of the app (which can be done by querying for its mime type in the BRoster, I think). But that solution will not work as well for the translators, where there is not really a main app to put the catalog into.

> Support is reused and contains quite a lot of keys that would have to be translated for both AddOnHost and RouteApp. I though about making it shared library, but if we could have APIs described above it wouldn't be needed.

Yes, let's just add that. shared libs in add-ons is asking for even more trouble :)

Do you want to make the changes to locale kit, or should I look into it?

Patch-set: 1